Adopted API changes
git-svn-id: https://svn.code.sf.net/p/jdbacl/code/trunk@529 ae0afc29-7cfe-489a-95cf-4bb878f6dcf4

+++ src/test/java/org/databene/jdbacl/dialect/DatabaseDialectTest.java
@@ -117,7 +117,7 @@ public abstract class DatabaseDialectTest<E extends DatabaseDialect> {
 			DBSequence[] sequences = dialect.querySequences(connection);
 			String[] sequenceNames = NameUtil.getNames(sequences);
 			assertTrue(StringUtil.containsIgnoreCase(sequenceName, sequenceNames));
-			assertEquals(23L, DBUtil.queryLong(dialect.renderFetchSequenceValue(sequenceName), connection));
+			assertEquals(23L, DBUtil.queryLong(dialect.renderFetchSequenceValue(sequenceName), connection).longValue());
 		} finally {
 			DBUtil.executeUpdate(dialect.renderDropSequence(sequenceName), connection);
 			DBUtil.close(connection);
